Move LLVMGPURemoveTrivialLoops to Common/ and use it for SPIR-V
This removes the ad-hoc trip one loop removal pass in the SPIR-V
backend; so better unification and less code to maintain! :)
