+++ iree/compiler/Codegen/Common/BUILD
@@ -41,6 +41,7 @@ cc_library(
         "IREEComprehensiveBufferizePass.cpp",
         "LinalgBufferizePass.cpp",
         "OptimizeVectorTransferPass.cpp",
+        "RemoveTrivialLoops.cpp",
         "SetNumWorkgroupsPass.cpp",
         "VectorizeConv.cpp",
         "VectorizeMMT4d.cpp",
