+++ iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -25,7 +25,6 @@ iree_cc_library(
     "LLVMGPUDistributeSharedMemoryCopy.cpp"
     "LLVMGPULowerExecutableTarget.cpp"
     "LLVMGPUPipelining.cpp"
-    "LLVMGPURemoveTrivialLoops.cpp"
     "LLVMGPUTileAndDistribute.cpp"
     "LLVMGPUVectorLowering.cpp"
     "LLVMGPUVectorization.cpp"
