+++ iree/compiler/Codegen/SPIRV/CMakeLists.txt
@@ -26,7 +26,6 @@ iree_cc_library(
     "Passes.cpp"
     "SPIRVCopyToWorkgroupMemory.cpp"
     "SPIRVLowerExecutableTargetPass.cpp"
-    "SPIRVRemoveOneTripTiledLoops.cpp"
     "SPIRVTileAndDistribute.cpp"
     "SPIRVTileAndVectorizeToCooperativeOps.cpp"
     "SPIRVVectorToCooperativeOps.cpp"
