+++ iree/compiler/Codegen/SPIRV/test/CMakeLists.txt
@@ -25,10 +25,9 @@ iree_lit_test_suite(
     "convert_to_spirv.mlir"
     "pipeline_matmul_cooperative_ops.mlir"
     "pipeline_matmul_vectorization.mlir"
-    "remove_one_trip_tiled_loop.mlir"
+    "tile_and_distribute.mlir"
     "tile_and_distribute_scatter.mlir"
     "tile_and_distribute_sort.mlir"
-    "tile_and_vectorize.mlir"
     "tile_and_vectorize_batch_matmul.mlir"
     "tile_and_vectorize_conv.mlir"
     "tile_and_vectorize_matmul.mlir"
