+++ iree/compiler/Codegen/SPIRV/test/tile_and_distribute_sort.mlir
@@ -55,10 +55,17 @@ hal.executable private @static_3d_sort  {
 //       CHECK:     %[[WG_INPUT_CAST:.+]] = memref.cast %[[WG_INPUT]]
 //       CHECK:     %[[WG_OUTPUT:.+]] = memref.subview %[[ARG1]]
 //       CHECK:     %[[TID_X:.+]] = "gpu.thread_id"() {dimension = "x"}
+//       CHECK:     %[[DIM_X:.+]] = "gpu.block_dim"() {dimension = "x"}
 //       CHECK:     %[[TID_Y:.+]] = "gpu.thread_id"() {dimension = "y"}
-//       CHECK:     %[[COPY_SOURCE:.+]] = memref.subview %[[WG_INPUT_CAST]][%[[TID_Y]], 0, %[[TID_X]]]
-//       CHECK:     %[[COPY_DEST:.+]] = memref.subview %[[WG_OUTPUT]][%[[TID_Y]], 0, %[[TID_X]]]
-//       CHECK:     linalg.copy(%[[COPY_SOURCE]], %[[COPY_DEST]])
-//       CHECK:     %[[T_OUTPUT_CAST:.+]] = memref.cast %[[COPY_DEST]]
-//       CHECK:     linalg_ext.sort dimension(1)
-//  CHECK-SAME:       outs(%[[T_OUTPUT_CAST]]
+//       CHECK:     %[[DIM_Y:.+]] = "gpu.block_dim"() {dimension = "y"}
+//       CHECK:     scf.for %[[IV_Y:.+]] = %[[TID_Y]] to %{{.+}} step %[[DIM_Y]]
+//       CHECK:       scf.for %[[IV_X:.+]] = %[[TID_X]] to %{{.+}} step %[[DIM_X]]
+//       CHECK:         %[[COPY_SOURCE:.+]] = memref.subview %[[WG_INPUT_CAST]][%[[IV_Y]], 0, %[[IV_X]]]
+//       CHECK:         %[[COPY_DEST:.+]] = memref.subview %[[WG_OUTPUT]][%[[IV_Y]], 0, %[[IV_X]]]
+//       CHECK:         linalg.copy(%[[COPY_SOURCE]], %[[COPY_DEST]])
+//       CHECK:     scf.for %[[IV_Y:.+]] = %[[TID_Y]] to %{{.+}} step %[[DIM_Y]]
+//       CHECK:       scf.for %[[IV_X:.+]] = %[[TID_X]] to %{{.+}} step %[[DIM_X]]
+//       CHECK:         %[[COPY_DEST:.+]] = memref.subview %[[WG_OUTPUT]][%[[IV_Y]], 0, %[[IV_X]]]
+//       CHECK:         %[[T_OUTPUT_CAST:.+]] = memref.cast %[[COPY_DEST]]
+//       CHECK:         linalg_ext.sort dimension(1)
+//  CHECK-SAME:           outs(%[[T_OUTPUT_CAST]]
